requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
phonenumbers==8.13.26